feat: enhance E2E tests by adding seed_bars fixture and improving database cleanup logic
