+++ tests/e2e/test_paper_flow.py
@@ -11,10 +11,9 @@ Tests focus on lifecycle state transitions only.
 
 from __future__ import annotations
 
-import contextlib
-
 import pytest
 from playwright.sync_api import Page, expect
+from requests import HTTPError
 
 from tests.e2e.helpers import E2EApiClient
 
@@ -91,7 +90,7 @@ class TestPaperFlow:
             mode="paper",
             symbols=["BTC/USD"],
         )
-        with contextlib.suppress(Exception):
+        with pytest.raises(HTTPError):
             api_client.start_run(run["id"])
 
         page.goto(f"{e2e_base_url}/runs")
